Add ROM loading functionality
Added the ability to load a buffer into CHIP-8 memory. Currently, the
ROM filepath is hard coded in main(), but I would eventually like to
add a GUI to select which ROM to run. I'll also probably move the file
loading code out of main() and into some other file or class sooner or
later.