Add support for virtual lookup table in core
In order to avoid the inefficiency (and ugliness) of
a massive switch statement to execute opcodes, I am working
on implementing a virtual lookup table by using the opcode
value to index into an array of member function pointers.

The current plan is to use the MSB of each opcode to index into the array.
Since there exist multiple opcodes for certain MSB values, further
processing may be necessary after the initial index.