Implement rest of opcodes
As far as I can tell, everything is working so far. The gfx array
is being updated on draw commands which is good. Next, I need to pass
in updates to key state on each emulation cycle as well as render the pixel
array through SDL. I also need to implement timing somehow.